scripts/requirements.txt
# Core browser automation
playwright>=1.50

# Image generation / validation
Pillow>=10.0
numpy>=1.26

# Optional desktop fallback automation
PyAutoGUI>=0.9

# YouTube OAuth + API automation
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
